Add support for MacOS .plist files
The files are not technically classified as dotfiles (not starting with
a dot). Only already collected dotfiles are used (i.e. present in
dotfiles/Libray/Preferences/ folder)

When collecting, the files are converted into XML for easier reading and
cleaned from not useful keys, e.g. time since last synchronization. The
list of excluded keys is in the config.

Config got upgraded from a custom file format into yaml. Yaml can
provide structured way how to add further knobs in the future.

The install script now depends on plutil when on Mac. As far as I can
tell, plutil is present on Macs by default.